Removes warning on verifying stubbed mocks
As TestDouble points out on the documentation[1], when writing a test
that uses a response from a stub, it is not necessary to verify it has
been called.

It is better instead to verify that the expected response has been used
as part of the expected response.

This commit removes the verification that was causing an Warning to be
printed.

[1]
https://github.com/testdouble/testdouble.js/blob/master/docs/B-frequently-asked-questions.md#why-shouldnt-i-call-both-tdwhen-and-tdverify-for-a-single-interaction-with-a-test-double