Recreates application and http client between requests
At this moment, we are creating a single http client that we are
registering several stub responses. Those stub responses are being
maintained across tests, which makes the next request dirty, with
previously registered endpoints.

This commit recreates the application and the http client mock before
each test.

To do the recreation, it was needed to change how we import the helper.
We were using the `--require test/helper.js` on the mocha CLI arguments,
but it lacks the capabilities to setup global hooks[1].

This forced us to move to the pattern to use `import './helper.js'` on
top of each file to allow a global reset hook.

[1] https://github.com/mochajs/mocha/issues/764