Fix Equipment API integration issue
In case Telemetry API is not available the Equipment API keeps the
connection with the client open forever without sending any response.
The same behavior is reproducible in case we provide an invalid
Telemetry API URL,
In order to avoid that, we should close connections as soon as we find
this kind of problem.