Adding timeouts to our HTTP requests
Modern systems integrate with a lot of external dependencies. Most of
these dependencies are outside our control. While requesting content
from one of them, it could happen that it does not reply our
requests. Even worse would be if it decides to answer our requests at a
rate of one byte per second!

What could we do in this case? That is where Timeouts comes to rescue!