Refactors code based on feedback
We got a few feedbacks from the team and we refactored our code to
comply with the suggestions.

On `src/metrics/transformers/userInfoTransformer.js` the use of
`Promise` is not mandatory. Given this operation can happen
synchronously we can remove it.

On `src/metrics/responseTimeExtractor.js` we could use `Object.assign`
to merge default information with user data.