Adds CORS support
We thought we had fixed CORS issues a while ago. But it seems that we
haven't.

Today I noticed that the Facade does not support CORS while executing
our dealership demo example.

In order to finally fix this I included a simple hapi plugin which
includes all the necessary CORS headers.