Fixing dependencies vulnerabilities
This morning we noticed that several npm packages are broken due to a
high severity vulnerability on minimatch package.

This problem started today. We updated our dependencies to their latest
version, but most of them haven't updated.

We set it to ignore this problem for the next 30 days. That might be
enough time to get updates on most of these packages.

At the same time, one of our dependencies was `mocha-bamboo-reporter`.
This package was no longer needed and we got rid of it.